Code organization and review suggestions
Changes (non-exhaustive):
* FloatOrd supports NaN
* BoxElem combines 5 values in BoxSpread struct
* Split bar/box-related functionality into multiple files
* Pos2::distance_from_rect_sq() -> Rect::distance_sq_to_pos()
* Consistent terminology (key -> argument)
