+++ egui/src/util/float_ord.rs
@@ -10,20 +10,34 @@ pub(crate) struct OrderedFloat(f32);
 impl Eq for OrderedFloat {}
 
 impl PartialEq<Self> for OrderedFloat {
+    #[inline]
     fn eq(&self, other: &Self) -> bool {
-        self.0 == other.0
+        // NaNs are considered equal (equivalent when it comes to ordering
+        if self.0.is_nan() {
+            other.0.is_nan()
+        } else {
+            self.0 == other.0
+        }
     }
 }
 
 impl PartialOrd<Self> for OrderedFloat {
+    #[inline]
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        self.0.partial_cmp(&other.0)
+        match self.0.partial_cmp(&other.0) {
+            Some(ord) => Some(ord),
+            None => Some(self.0.is_nan().cmp(&other.0.is_nan())),
+        }
     }
 }
 
 impl Ord for OrderedFloat {
+    #[inline]
     fn cmp(&self, other: &Self) -> Ordering {
-        self.partial_cmp(other).expect("NaN or Inf not permitted")
+        match self.partial_cmp(other) {
+            Some(ord) => ord,
+            None => unreachable!(),
+        }
     }
 }
 
@@ -34,6 +48,7 @@ pub(crate) trait FloatOrd {
 }
 
 impl FloatOrd for f32 {
+    #[inline]
     fn ord(self) -> OrderedFloat {
         OrderedFloat(self)
     }
